Move getEarliestFutureDateRange alongside its component; add a comment
